Getting close to Scrimmage-Ready.
WARNING CARL: if you have been using any of my code, it won't quite work right for you right now. I changed the hardwaremap to fix the front of the robot being on the wrong side and I have not updated the wiring guide or uploaded the xml.

Lost of things, but mostly just squishing things together into the two PreProduction programs to make programs for the Scrimmage.